re-inject the best micro evolution into a big pool of trees for a macro evolution. this is great since it is going to concatenate some bills and we are going to be able to make a list of predictors now. going to bed

+++ core/acdata.h
@@ -177,7 +177,7 @@ public:
 		}
 		return ret;
 	}
-	double averageKLA() const {
+	int averageKLA() const {
 		double ret = sumDollar() / m_vector.count();
 		return kindaLog(ret) * KLA_MULTIPLICATOR;
 	}
